added unique ids, added the ability to delete events
