+++ client/package.json
@@ -26,6 +26,7 @@
     "react-scripts": "^5.0.1",
     "react-transition-group": "4.4.2",
     "react-window": "1.8.7",
+    "uuid": "^8.3.2",
     "web-vitals": "2.1.4",
     "workbox-background-sync": "6.5.3",
     "workbox-broadcast-update": "6.5.3",
@@ -71,6 +72,7 @@
     "@types/react-router-dom": "5.3.3",
     "@types/react-transition-group": "4.4.4",
     "@types/react-window": "1.8.5",
+    "@types/uuid": "^8.3.4",
     "prettier": "2.6.2",
     "typescript": "4.6.4"
   }
